вынес утилиты, переписал клавиатуру и исправил импорты
