+++ handlers/user_handlers.py
@@ -5,9 +5,10 @@ from aiogram.filters import Command, BaseFilter
 from aiogram.exceptions import TelegramBadRequest
 
 from create_bot import bot
-from file_managment.ya_file_manager import upload_img, get_directories_in
-from file_managment.file_moving_manager import get_dt_name, move_to_folders_on_disk, path_list, set_correct_path
+from file_managment.ya_file_manager import upload_img, get_directories_in, path_list
+from file_managment.file_moving_manager import move_to_folders_on_disk
 from keyboards.user_keyboards import get_gir_keyboard, system_buttons
+from utils.utils import get_dt_name, set_correct_path
 
 
 router = Router()
@@ -24,9 +25,7 @@ async def start_message(message: types.message):
     path_list.clear()
     await message.answer(
         text='Бот для сохранения пересылаемых изображений в облако. Как в папке выбираешь директорию в которую желаешь поместить файл и отправляешь репост или фото')
-    await message.answer('Выбери директорию:',
-                         reply_markup=get_gir_keyboard(buttons(), path_list).as_markup(resize_keyboard=True)
-                         )
+    await message.answer('Выбери директорию:', reply_markup=get_gir_keyboard())
 
 
 class FolderFilter(BaseFilter):
@@ -46,14 +45,16 @@ async def move_to_dir(message: types.Message, bot: bot):
     await message.answer('Идет запрос к директориям диска...')
     move_to_folders_on_disk(message.text)
     await message.answer(set_correct_path(path_list) if len(path_list) else 'Корневая директория/',
-                         reply_markup=get_gir_keyboard(buttons(), path_list).as_markup(resize_keyboard=True))
-    mes_id = [message.message_id + 1, message.message_id, message.message_id - 1]
+                         reply_markup=get_gir_keyboard())
+    mes_id_list = [message.message_id + 1, message.message_id, message.message_id - 1]
     try:
-        for i in mes_id:
+        for i in mes_id_list:
             await bot.delete_message(message.from_user.id, i)
     except TelegramBadRequest as ex:
         if ex.message == 'Bad Request: message to delete not found':
             print('Список id пуст')
+    print(path_list)
+    print(set_correct_path(path_list))
 
 
 @router.message(F.photo)  # Хендлер на присланные фотографии
