+++ keyboards/user_keyboards.py
@@ -1,6 +1,9 @@
 from aiogram.types import KeyboardButton
 from aiogram.utils.keyboard import ReplyKeyboardBuilder
 
+from file_managment.ya_file_manager import path_list
+from file_managment.ya_file_manager import get_directories_in
+
 
 system_buttons = {  # Кнопки возврата в предыдущую директорию или в корневую
     'home': '🏠',
@@ -8,12 +11,11 @@ system_buttons = {  # Кнопки возврата в предыдущую ди
     }
 
 
-def get_gir_keyboard(buttons, path_list):
+def get_gir_keyboard():
     """Компоновка клавиатуры из видимых директорий"""
     builder = ReplyKeyboardBuilder()
     if len(path_list):
-        [builder.add(KeyboardButton(text=el)) for el in buttons]
-    else:
-        [builder.add(KeyboardButton(text=el)) for el in buttons[2:]]
+        [builder.add(KeyboardButton(text=el)) for el in system_buttons.values()]
+    [builder.add(KeyboardButton(text=el)) for el in get_directories_in(path_list)]
     builder.adjust(2)
-    return builder
+    return builder.as_markup(resize_keyboard=True)
